fix: Update ACRA to correctly build & use email
After updating gradle (etc), the project fails to build with
assembleBetaRelease, because of ACRA.
This commit updates ACRA to make the build work.
It also uses email as the method, instead of external files, to be more
robust.
Also removed proguar-rules-acra.pro, because it is unnecessary -- ACRA
has its own proguard rules which will be automatically picked up.